server/requirements-freeze.txt
annotated-types==0.7.0
astral==3.2
astropy==7.1.0
astropy-iers-data==0.2025.9.22.0.37.25
attrs==24.3.0
blinker==1.9.0
boolean.py==5.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==44.0.3
cyclonedx-python-lib==9.1.0
defusedxml==0.7.1
Deprecated==1.2.18
filelock==3.20.0
Flask==3.1.1
Flask-Cors==4.0.0
google-api-core==2.28.1
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.71.0
grpcio==1.71.0
grpcio-status==1.62.3
h11==0.14.0
httplib2==0.22.0
idna==3.10
importlib-metadata==7.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
license-expression==30.4.4
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.0
mysql-connector-python==9.1.0
numpy==2.3.3
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
outcome==1.3.0.post0
packageurl-python==0.17.5
packaging==25.0
pillow==11.0.0
pip-api==0.0.34
pip-licenses==5.5.0
pip-requirements-parser==32.0.1
pip_audit==2.9.0
platformdirs==4.5.0
prettytable==3.16.0
proto-plus==1.26.1
protobuf==4.25.8
py-serializable==2.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pyerfa==2.0.1.5
pygame==2.6.0
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
PySocks==1.7.1
python-dotenv==1.1.1
pytz==2024.2
PyYAML==6.0.1
requests==2.32.4
rich==14.2.0
rsa==4.9.1
selenium==4.28.1
sniffio==1.3.1
sortedcontainers==2.4.0
tenacity==9.0.0
toml==0.10.2
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.5.0
wcwidth==0.2.14
websocket-client==1.8.0
Werkzeug==3.1.3
WooCommerce==3.0.0
wrapt==1.17.3
wsproto==1.2.0
zipp==3.23.0
